smart-financial-assitant/requirements.txt
langchain
langgraph
streamlit
groq
langchain-groq
langchain-community
python-dotenv